pyproject.toml
[tool.poetry]
name = "pandasai"
version = "3.0.0"
description = "Chat with your database (SQL, CSV, pandas, mongodb, noSQL, etc). PandasAI makes data analysis conversational using LLMs (GPT 3.5 / 4, Anthropic, VertexAI) and RAG."
authors = ["Gabriele Venturi"]
license = "MIT"
readme = "README.md"
packages = [{include = "pandasai"}]

[tool.poetry.urls]
"Documentation" = "https://docs.pandas-ai.com/"
"Repository" = "https://github.com/sinaptik-ai/pandas-ai"

[tool.poetry.dependencies]
python = ">=3.8"
python-dotenv = "^1.0.0"
pandas = "^2.0.3"
scipy = ">=1.10.1"
astor = "^0.8.1"
matplotlib = ">=3.7.1"
pydantic = "^2.6.4"
duckdb = "^1.0.0"
pillow = "^10.1.0"
requests = "^2.31.0"
jinja2 = "^3.1.3"
numpy = "^1.17"
openpyxl = "^3.1.5"
seaborn = "^0.12.2"
sqlglot = "^25.0.3" 
pyarrow = ">=14.0.1"
pyyaml = "^6.0.2"
python-pptx = "^1.0.2"
reportlab = "<4.0"
pyjwt = "2.8.0"
sqlalchemy = "^2.0.48"
passlib = "^1.7.4"
bcrypt = "^5.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
ruff = "^0.1.0"
codespell = "^2.2.0"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
pytest-env = "^0.8.1"
click = "^8.1.3"
coverage = "^7.2.7"
sourcery = "^1.11.0"
openai = "^1.60.0"

[tool.poetry.scripts]
pai = "pandasai.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["tests_*"]

[tool.setuptools]
license-files = ["LICENSE"]